version=11.9.21_dev  rapid prototyping STRUTS
# new features for the Center for Biomolecular Modeling, Milwaukee School of Engineering
#              contributed ideas of Mark Hoeltzer for rapid prototyping.
#
# new feature: STRUTS shape -- for rapid prototyping -- these are sticks that 
#              can be connected anywhere but are not measures and are not covalent bonds
#
# new feature: calculate STRUTS  -- contributed algorithm of George Phillips
#              (phillips@biochem.wisc.edu) automates introduction of struts
#
# new feature: set picking CONNECT  -- like measuring distances, except you get bonds.
# new feature: set picking DELETEBONDS -- opposite idea -- click two atoms generally,
#                                         or, with set BONDPICKING, click the center of the bond
#
# new feature: set strutSpacing 6  -- George's "DELTA" 
# new feature: set strutLengthMaximum 7.0 -- George's "THRESH" 
# new feature: set strutDefaultRadius 0.3 -- default rendering (will be translucent white as well)

+++ Jmol/src/org/jmol/api/JmolViewer.java
@@ -340,7 +340,7 @@ abstract public class JmolViewer extends JmolSimpleViewer {
   abstract public Point3f getBondPoint3f2(int bondIndex);
   abstract public int getBondArgb1(int bondIndex);
   abstract public int getBondArgb2(int bondIndex);
-  abstract public short getBondOrder(int bondIndex);
+  abstract public int getBondOrder(int bondIndex);
   abstract public int getBondModelIndex(int bondIndex);
 
   abstract public Point3f[] getPolymerLeadMidPoints(int modelIndex, int polymerIndex);
